extract CompetencyHeader and ScoreLevelSelector from AssessmentForm
